make a new system to find time
okay, so this one doesn't suck anymore.
in the past, since it was just a simple regex, it would sometimes pull
time out of links, breaking the link and making the time longer than
the user wanted it to be. this new system is more of a parser, and
doesn't have that issue whatsoever.